'get_docmap_evaluation_type_form_tags' has been moved into new module
