+++ data_hub_api/docmaps/codecs/evaluation.py
@@ -4,6 +4,10 @@ from data_hub_api.docmaps.codecs.elife_manuscript import get_elife_manuscript_ve
 
 DOI_ROOT_URL = 'https://doi.org/'
 
+DOCMAP_EVALUATION_TYPE_FOR_EVALUATION_SUMMARY = 'evaluation-summary'
+DOCMAP_EVALUATION_TYPE_FOR_REPLY = 'reply'
+DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE = 'review-article'
+
 
 def get_elife_evaluation_doi(
     elife_doi_version_str: str,
@@ -27,3 +31,26 @@ def get_elife_evaluation_doi_url(
     if not elife_evaluation_doi:
         return None
     return f'{DOI_ROOT_URL}' + elife_evaluation_doi
+
+
+def has_tag_containing(tags: list, text: str) -> bool:
+    return any(
+        text in tag
+        for tag in tags
+    )
+
+
+def get_docmap_evaluation_type_form_tags(
+    tags: list
+) -> Optional[str]:
+    has_author_response_tag = has_tag_containing(tags, 'AuthorResponse')
+    has_summary_tag = has_tag_containing(tags, 'Summary')
+    has_review_tag = has_tag_containing(tags, 'Review')
+    assert not (has_author_response_tag and has_summary_tag)
+    if has_author_response_tag:
+        return DOCMAP_EVALUATION_TYPE_FOR_REPLY
+    if has_summary_tag:
+        return DOCMAP_EVALUATION_TYPE_FOR_EVALUATION_SUMMARY
+    if has_review_tag:
+        return DOCMAP_EVALUATION_TYPE_FOR_REVIEW_ARTICLE
+    return None
